drivers: display: esp32_dsi: widen the pixel format storage
The pixel format from the devicetree was stored in a uint8_t, but
display_pixel_format values are bit flags and everything past
PIXEL_FORMAT_AL_88 (BIT(7)) was silently truncated to zero. Declaring
e.g. PIXEL_FORMAT_BGR_888 (BIT(9)) on the controller node then made
the bits-per-pixel check refuse every panel.

+++ drivers/display/display_esp32_dsi.c
@@ -35,7 +35,7 @@ struct display_esp32_dsi_config {
 	uint8_t irq_flags;
 	uint16_t width;
 	uint16_t height;
-	uint8_t pixel_format;
+	uint32_t pixel_format;
 };
 
 struct display_esp32_dsi_data {
